// swiyu-ecosystem-portal-ui/src/app/api/generated/model/business-partner-trust-status.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const BusinessPartnerTrustStatus = {
  NotVerified: 'NOT_VERIFIED',
  VerificationStarted: 'VERIFICATION_STARTED',
  VerificationInProgress: 'VERIFICATION_IN_PROGRESS',
  InformationRequested: 'INFORMATION_REQUESTED',
  Verified: 'VERIFIED',
  ReVerificationStarted: 'RE_VERIFICATION_STARTED',
  ReVerificationInProgress: 'RE_VERIFICATION_IN_PROGRESS'
} as const;
export type BusinessPartnerTrustStatus = (typeof BusinessPartnerTrustStatus)[keyof typeof BusinessPartnerTrustStatus];